+++ Currency-exchange/src/main/java/ru/viktorgezz/util/converter/ExchangeConverter.java
@@ -4,8 +4,10 @@ import ru.viktorgezz.dto.ExchangeDto;
 import ru.viktorgezz.model.Exchange;
 import ru.viktorgezz.model.ExchangeRate;
 
+import java.math.BigDecimal;
+
 public class ExchangeConverter {
-    public static ExchangeDto convertPartsOfModelToDto(ExchangeRate exchangeRate, double amount,double convertedAmount) {
+    public static ExchangeDto convertPartsOfModelToDto(ExchangeRate exchangeRate, BigDecimal amount, BigDecimal convertedAmount) {
         return new ExchangeDto.Builder()
                 .setExchangeRate(exchangeRate)
                 .setAmount(amount)
